Update it_it.json (#50)
Translation completed. 
I've kept the phrases said to the player as gender neutral as possible. 
Also kept the same emotions' intensity as close as possible from what it is meant to impart to the player.
I have to say that some terminology in english doesn't exist in a one-word term in Italian, so i had to make some phrases longer than they should just to make sure they had the same exact meaning and feeling in it.